# gltfpack: Optimize .obj parsing for memory consumption