@zeux 2022-07-29
Looking closer at rungholt, I'm not overly concerned with this as the gain doesn't show up when measuring allocation usage, so it's highly allocator specific, might be related to fragmentation. Also that was on the file without .mtl loaded, with .mtl the baseline memory consumption is higher and the relative increase is lower. Overall this change seems important and beneficial, we can investigate other memory reductions separately.