@zeux 2022-07-31
I've double checked and the increase is indeed specific to the system allocator used; I don't think it's fragmentation-related and instead must be related to the policy for freeing pages to the OS. Using mimalloc as the allocator actually reverses the results, with the peak RSS being *lower* after this change.