@zeux 2022-07-31
A few other large files show improvement as expected, eg powerplant.obj:

Before: 6.56 sec 2575132 KB
After: 5.49 sec 1250440 KB